requirements.txt
opencv-python
pillow
matplotlib
scipy
